Fixing color scheme
This adds new Sass variables, `$primary` and `$dsa-red`, and makes the former point to the latter. (This should make it cleaner to make changes in the future: now you know which color values are, say, green because green is  the "primary theme color" vs green because we want green to signal "success").

This depends on #16 because it also incidentally fixes a small inconsistency with the uses of logos across layouts (the "post" layout, unlike the others, did not use the 50x50 logo for small logos).

Fixes #14.